Various fixes:
Found out that setup.sh wasn't iterating over the list of file names.
Only .zshrc was being added to /home/drfiresign. Corrected so .tmux.conf and
.vimrc are symlinked as well. I also corrected a spacing error that
was appearing on tmux in the left status bar.

If only I could figure out how to actually get vim to look at that
configuration file... It's not reading the set number command and
I am so confused by that.